+++ app/dashboard/layout.tsx
@@ -1,18 +1,19 @@
 import { AppNavbar } from "@/components/ui/nav/app.navbar";
-import { SidebarInset } from "@/components/ui/sidebar";
+import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 import { AppSidebar } from "@/components/ui/sidebar/root-sidebar";
 import { ChildNodeProps } from "@/lib/interfaces/interface";
 import { FC } from "react";
 
 const layout: FC<ChildNodeProps> = ({ children }) => {
     return (
-        <>
+        <SidebarProvider>
             <AppSidebar />
             <SidebarInset>
-                <AppNavbar />
-                {children}
+                <header className="relative">
+                    <AppNavbar />
+                </header>
             </SidebarInset>
-        </>
+        </SidebarProvider>
     );
 };
