+++ app/globals.css
@@ -43,6 +43,7 @@
     --radius-md: calc(var(--radius) - 2px);
     --radius-lg: var(--radius);
     --radius-xl: calc(var(--radius) + 4px);
+    --sidebar-width: var(--sidebar-width, 16rem);
 }
 
 :root {
@@ -80,6 +81,7 @@
     --sidebar-accent-foreground: oklch(0.205 0 0);
     --sidebar-border: oklch(0.922 0 0);
     --sidebar-ring: oklch(0.708 0 0);
+    --sidebar-width: 16rem;
 }
 
 .dark {
@@ -116,6 +118,7 @@
     --sidebar-accent-foreground: oklch(0.985 0 0);
     --sidebar-border: oklch(1 0 0 / 10%);
     --sidebar-ring: oklch(0.556 0 0);
+    --sidebar-width: 16rem;
 }
 
 @layer base {
